% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algor.R
\name{cmatches}
\alias{cmatches}
\title{one<-as.integer(1:1000000)
two<-as.integer(sample(1:1000000,1e5,TRUE))
system.time(a<-lapply(one, function (x) which(two %in% x)))
system.time(b<-cmatches(one,two,type='full',list=TRUE))}
\usage{
cmatches(a, b, type = "left", list = FALSE)
}
\description{
one<-as.integer(1:1000000)
two<-as.integer(sample(1:1000000,1e5,TRUE))
system.time(a<-lapply(one, function (x) which(two %in% x)))
system.time(b<-cmatches(one,two,type='full',list=TRUE))
}

