% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algor.R
\name{corder}
\alias{corder}
\title{chars<-as.character(sample(1e3,1e4,TRUE))
system.time(a<-order(chars))
system.time(b<-corder(chars))}
\usage{
corder(x)
}
\description{
chars<-as.character(sample(1e3,1e4,TRUE))
system.time(a<-order(chars))
system.time(b<-corder(chars))
}

