% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algor.R
\name{csort}
\alias{csort}
\title{chars<-as.character(sample(1e3,1e4,TRUE))
system.time(a<-sort(chars))
system.time(b<-csort(chars))}
\usage{
csort(x)
}
\description{
chars<-as.character(sample(1e3,1e4,TRUE))
system.time(a<-sort(chars))
system.time(b<-csort(chars))
}

