% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algor.R
\docType{package}
\name{algor}
\alias{algor}
\alias{algor-package}
\title{Faster implementations of base R functions, including sort, order, and match.}
\description{
Faster implementations of base R functions, including sort, order, and match.
}

